+++ lib/Atlas/src/Reference.cpp
@@ -1,10 +1,10 @@
 #include "Reference.h"
 
-Reference::Reference(std::string* name, double (*callback)(void)): Value(name) {
+Reference::Reference(std::string* name, float (*callback)(void)): Value(name) {
     this->callback = callback;
 }
 
-Reference::Reference(double (*callback)(void)): Reference(nullptr, callback) {
+Reference::Reference(float (*callback)(void)): Reference(nullptr, callback) {
 
 }
 
@@ -12,6 +12,6 @@ bool Reference::isStatic() {
     return false;
 }
 
-double Reference::get() {
+float Reference::get() {
     return callback();
 }
