+++ lib/Atlas/src/Variable.cpp
@@ -1,11 +1,11 @@
 #include "Variable.h"
 #include <string>
 
-Variable::Variable(std::string* name, double value): Value(name) {
+Variable::Variable(std::string* name, float value): Value(name) {
     this->value = value;
 }
 
-Variable::Variable(double value): Value() {
+Variable::Variable(float value): Value() {
     this->value = value;
 }
 
@@ -13,6 +13,6 @@ bool Variable::isStatic() {
     return true;
 }
 
-double Variable::get() {
+float Variable::get() {
     return this->value;
 }
